% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_code.R
\name{add.attribute.column}
\alias{add.attribute.column}
\title{add a resource attribute column}
\usage{
add.attribute.column(
  config,
  column.name,
  label = "",
  description = "",
  from.existing = "",
  definition.uri = "",
  values = "",
  required = "",
  from.template = ""
)
}
\arguments{
\item{config}{A configuration object. The column configuration will be added to the configuration object.}

\item{column.name}{String. A column name in the dataframe that the configuration object corresponds.}

\item{label}{String. Label describing the Attribute.}

\item{description}{String. A explanation of what the Attribute represents.}

\item{from.existing}{String. URI of an existing attribute to reuse or extend. Can use the built-in list "Attibutes" to auto-fill.}

\item{definition.uri}{String. Source URI for the definition of the Attribute.}

\item{values}{Can  be:
boolean: Whether to automatically generate attribute value resources for each unique value in the column.
A "values" object. Created by the values() function filled with attribute.value(). The distinct attribute values which can be used in the column}

\item{required}{Boolean. Whether or not it is required that every observation has a value for this attribute.}

\item{from.template}{String. Attribute template uris.}
}
\value{
A configuration object with the new column added to it.
}
\description{
add a resource type attribute column to be included in the configuration object.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.attribute.column(config,
  column.name = "Status", label = "status",
  values = values(attribute.value(label = "Final"), attribute.value(label = "Provisional"))
)

}
