% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_code.R
\name{add.dimension.column}
\alias{add.dimension.column}
\title{Add a dimension attribution column}
\usage{
add.dimension.column(
  config,
  column.name = "",
  label = "",
  description = "",
  from.existing = "",
  definition.uri = "",
  cell.uri.template = "",
  code.list = "",
  from.template = ""
)
}
\arguments{
\item{config}{A configuration object.The configuration object to add the column to.}

\item{column.name}{String. A column name in the Data Frame.}

\item{label}{String. Label describing the Dimension.}

\item{description}{String. A explanation of what the Dimension represents.}

\item{from.existing}{String. URI of an existing dimension to reuse or extend. Use built-in list dimension.from.existing to auto-fill.}

\item{definition.uri}{String. Source URI for the definition of the Dimension}

\item{cell.uri.template}{String. A template used to map the cell values in this column into URIs, e.g. 'http://example.com/code-list/{+column_csvw_name}'}

\item{code.list}{"A code list defines the unique values permitted in a dimension"
Boolean: Whether or not to generate a code list defining the unique values permitted in this dimension.
String:The URI of a code list defining the unique values permitted in this dimension
String: The path to the code list json defined using the code list config schema.}

\item{from.template}{String. Dimension column templates. Use cuilt-in list dimension.from.template to autofill. Also, Users can use add.year.column() and other functions to achieve the same effect.}
}
\value{
A configuration object with the new column added to it.
}
\description{
Add a dimension column to the configuration.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.dimension.column(config, column.name = "Year",
   from.template = dimension.from.template$year,
   from.existing = dimension.from.existing$timePeriod)
}
