% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_template.R
\name{add.governmentyear.column}
\alias{add.governmentyear.column}
\title{Add a government year dimension attribution column}
\usage{
add.governmentyear.column(
  config,
  column.name,
  label = "",
  description = "",
  from.existing = "",
  definition.uri = "",
  cell.uri.template = "",
  code.list = ""
)
}
\arguments{
\item{config}{A configuration ob.The configuration object to add the column to.}

\item{column.name}{String. A column name in the Data Frame.}

\item{label}{String. Label describing the Dimension.}

\item{description}{String. A explanation of what the Dimension represents.}

\item{from.existing}{String. URI of an existing dimension to reuse or extend. Use list dimension.from.existing to auto-fill.}

\item{definition.uri}{String. Source URI for the definition of the Dimension}

\item{cell.uri.template}{String. A template used to map the cell values in this column into URIs, e.g. 'http://example.com/code-list/{+column_csvw_name}'}

\item{code.list}{"A code list defines the unique values permitted in a dimension"
Boolean: Whether or not to generate a code list defining the unique values permitted in this dimension.
String:The URI of a code list defining the unique values permitted in this dimension
String: The path to the code list json defined using the code list config schema.}
}
\value{
A configuration object with the new government year column added to it.
}
\description{
Add a government year dimension column to the configuration.
}
