% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_code.R
\name{add.literal.attribute.column}
\alias{add.literal.attribute.column}
\title{add a literal attribution column}
\usage{
add.literal.attribute.column(
  config,
  column.name = "",
  label = "",
  description = "",
  from.existing = "",
  definition.uri = "",
  data.type = "",
  required = ""
)
}
\arguments{
\item{config}{A configuration object. The column configuration will be added to the configuration object.}

\item{column.name}{String. A column name in the dataframe that the configuration object corresponds.}

\item{label}{String. Label describing the Attribute.}

\item{description}{String. A explanation of what the Attribute represents.}

\item{from.existing}{String. URI of an existing attribute to reuse or extend. Can use the built-in list "attribute.from.existing" to auto-fill.}

\item{definition.uri}{String. Source URI for the definition of the Attribute.}

\item{data.type}{String. "The data type of the attribute values"}

\item{required}{Boolean. Whether or not it is required that every observation has a value for this attribute.}
}
\value{
A configuration object with the new column added to it.
}
\description{
add a literal type attribute column to be included in the configuration object.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.literal.attribute.column(config, column.name = "Status", label = "status")
}
