% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_code.R
\name{add.measure.column}
\alias{add.measure.column}
\title{Add a measure column.}
\usage{
add.measure.column(config, column.name = "", values = "")
}
\arguments{
\item{config}{A configuration object.The configuration object to add the column to.}

\item{column.name}{String. A column name in the Data Frame.}

\item{values}{Can be:
boolean: Whether to automatically measure resources for each unique value in the column.
A values object. A list of measures permitted in this column. Created by the values() function filled with measure.value().}
}
\value{
the new configuration object with the measure column added.
}
\description{
Add a measure column to the configuration.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.measure.column(config, column.name = "Measure",
   values = values(measure.value(label = "Final Rank"),measure.value(label = "Final Points"),
   measure.value(label = "People on Stage")))
}
