% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_code.R
\name{add.unit.column}
\alias{add.unit.column}
\title{Add a unit column.}
\usage{
add.unit.column(config, column.name, unit.values = "")
}
\arguments{
\item{config}{A configuration object.The configuration object to add the column to.}

\item{column.name}{String. A column name in the Data Frame.}

\item{unit.values}{The Units used in this dataset.Can be:
A boolean. Whether to automatically generate new units from the unique values in this column.
A "unit.values" object. A list of units permitted in this column. Use unit.values() and unit() to build such an object.}
}
\description{
Add a unit column to the configuration.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
   unit.values = unit.values(unit(label = "Unitless"),
   unit(label = "Number"))
}
