% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_from_existing.R
\docType{data}
\name{attribute.from.existing}
\alias{attribute.from.existing}
\title{A list of URIs that can be used to fill in the from.existing argument in add.resource.attribute.column().}
\format{
An object of class \code{list} of length 54.
}
\usage{
attribute.from.existing
}
\description{
A list of URIs that can be used to fill in the from.existing argument in add.resource.attribute.column().
}
\keyword{datasets}
