% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_code.R
\name{attribute.value}
\alias{attribute.value}
\title{create a value object for filling in the values() function}
\usage{
attribute.value(
  label,
  description = "",
  definition.uri = "",
  from.existing = ""
)
}
\arguments{
\item{label}{String.A label describing this attribute value}

\item{description}{String. The description of the attribute values}

\item{definition.uri}{String. The URI link to an existing definition.}

\item{from.existing}{String. The URI to an existing attribute values to extend. Users can use the list AttibuteValues for auto-filling.}
}
\value{
a value object storing the attribute value information.
}
\description{
create one attribute value object for the value argument in the add.resource.attribute.column() function.
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.attribute.column(config,
  column.name = "Status", label = "status",
  values = values(attribute.value(label = "Final"), attribute.value(label = "Provisional"))
)
}
