% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_from_existing.R
\docType{data}
\name{Attributes}
\alias{Attributes}
\title{A list of URIs that can be used to fill in the from.existing argument in add.resource.attribute.column().}
\format{
An object of class \code{list} of length 43.
}
\usage{
Attributes
}
\description{
A list of URIs that can be used to fill in the from.existing argument in add.resource.attribute.column().
}
\keyword{datasets}
