% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_value_from_existing.R
\docType{data}
\name{AttributeValues}
\alias{AttributeValues}
\title{A list to be used to fill in the from.existing argument of attribute.value().}
\format{
An object of class \code{list} of length 10.
}
\usage{
AttributeValues
}
\description{
A list to be used to fill in the from.existing argument of attribute.value().
}
\keyword{datasets}
