% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\docType{data}
\name{DataTypes}
\alias{DataTypes}
\title{A list of strings that can be used to fill in the data.type argument in add.literal.attribute.column().}
\format{
An object of class \code{list} of length 22.
}
\usage{
DataTypes
}
\description{
A list of strings that can be used to fill in the data.type argument in add.literal.attribute.column().
}
\keyword{datasets}
