% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_from_template.R
\docType{data}
\name{dimension.from.template}
\alias{dimension.from.template}
\title{A list of strings that can be used to fill in the from.existing argument in add.resource.attribute.column().}
\format{
An object of class \code{list} of length 14.
}
\usage{
dimension.from.template
}
\description{
A list of strings that can be used to fill in the from.existing argument in add.resource.attribute.column().
}
\keyword{datasets}
