% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_from_existing.R
\docType{data}
\name{Dimensions}
\alias{Dimensions}
\title{A list of URIs that can be used to fill in the from.existing argument in add.dimension.column().}
\format{
An object of class \code{list} of length 10.
}
\usage{
Dimensions
}
\description{
A list of URIs that can be used to fill in the from.existing argument in add.dimension.column().
}
\keyword{datasets}
