% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_from_template.R
\docType{data}
\name{DimensionTemplates}
\alias{DimensionTemplates}
\title{A list of strings that can be used to fill in the from.template argument in add.dimension.column().}
\format{
An object of class \code{list} of length 14.
}
\usage{
DimensionTemplates
}
\description{
A list of strings that can be used to fill in the from.template argument in add.dimension.column().
}
\keyword{datasets}
