% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{sweden_at_eurovision_no_missing}
\format{
A data frame with 178 rows and 7 variables:
\describe{
  \item{Year}{Year}
  \item{Entrant}{name of entrant}
  \item{Song}{name of language}
  \item{Language}{Language of the song}
  \item{Value}{The value of the corresponding measure}
  \item{Measure}{the type of measure}
  \item{Unit}{the type of unit}
  \item{Status}{the status of result}
}
}
\source{
\url{https://raw.githubusercontent.com/GSS-Cogs/csvcubed-demo/v1.0/sweden_at_eurovision_no_missing.csv}
}
\usage{
example.data
}
\description{
An example data set to run examples.
}
\keyword{datasets}
