% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generate.json.configuration}
\alias{generate.json.configuration}
\title{Generate the configuration json file to the connection diretory. Returns the json string.}
\usage{
generate.json.configuration(config, file)
}
\arguments{
\item{config}{An configuration object. The configuration to be serialized into json.}

\item{file}{A string. The directory path and file name of the json file.}
}
\value{
The json scheme string of the  configuration.
}
\description{
Generate the configuration json file to the connection diretory. Returns the json string.
}
\examples{
data(example.data)
config <- create.config(df = example.data,
   id = "12341241", title = "sweden_at_eurovision_no_missing",
   dataset.issued = "2022-04-08",
   keywords = c("Eurovision", "Song Contest", "Sweden", "European Broadcasting Union"))
# the following will return the json string and generate a json file
# called "config.json" in the current working directory.
json.string <- generate.json.configuration(config, "config.json")

#remove the file
file.remove("config.json")
}
