% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licences.R
\docType{data}
\name{licences}
\alias{licences}
\title{A list of licence uris that could be used to fill in the license function argument in create.config().}
\format{
An object of class \code{list} of length 41.
}
\usage{
licences
}
\description{
A list of licence uris that could be used to fill in the license function argument in create.config().
}
\keyword{datasets}
