% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_value_from_existing.R
\docType{data}
\name{Measures}
\alias{Measures}
\title{A list of strings that can be used to fill in the from.existing argument in add.measure.column().}
\format{
An object of class \code{list} of length 11.
}
\usage{
Measures
}
\description{
A list of strings that can be used to fill in the from.existing argument in add.measure.column().
}
\keyword{datasets}
