% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\docType{data}
\name{Organizations}
\alias{Organizations}
\title{A list of organizations uris that could be used to fill in the creator and publisher function argument.}
\format{
An object of class \code{list} of length 1154.
}
\usage{
Organizations
}
\description{
A list of organizations uris that could be used to fill in the creator and publisher function argument.
}
\keyword{datasets}
