% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_kind.R
\docType{data}
\name{quantity.kind}
\alias{quantity.kind}
\title{A list of strings that can be used to fill in the quantity.kind argument in unit().}
\format{
An object of class \code{list} of length 861.
}
\usage{
quantity.kind
}
\description{
A list of strings that can be used to fill in the quantity.kind argument in unit().
}
\keyword{datasets}
