% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_kind.R
\docType{data}
\name{QuantityKinds}
\alias{QuantityKinds}
\title{A list of strings that can be used to fill in the quantity.kind argument in unit().}
\format{
An object of class \code{list} of length 861.
}
\usage{
QuantityKinds
}
\description{
A list of strings that can be used to fill in the quantity.kind argument in unit().
}
\keyword{datasets}
