% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_code.R
\name{unit}
\alias{unit}
\title{Create a unit object.}
\usage{
unit(
  label,
  description = "",
  from.existing = "",
  definition.uri = "",
  scaling.factor = "",
  quantity.kind = "",
  si.scaling.factor = ""
)
}
\arguments{
\item{label}{A string. Label describing the Unit.}

\item{description}{A string. A explanation of what the Unit represents.}

\item{from.existing}{A string. URI of an existing Unit to reuse. Can use list unit.value.from.existing to auto-fill.}

\item{definition.uri}{A string. Source URI for the definition of the Unit}

\item{scaling.factor}{A number. Scaling factor of the observation to base unit.}

\item{quantity.kind}{A string. The QUDT quantity kind this unit is used to measure.}

\item{si.scaling.factor}{A number. Scaling factor to convert from this unit to the SI base unit defined in the unit's quantity_kind.}
}
\description{
Create a unit object for add.unit.column() or add.observation.column()
}
\examples{
unit(label = "meter", description = "meter",
   from.existing = unit.value.from.existing$M, si.scaling.factor = 1,
   quantity.kind = quantity.kind$Length)
}
