% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_code.R
\name{unit.value}
\alias{unit.value}
\title{Create a unit object.}
\usage{
unit.value(
  label,
  description = "",
  from.existing = "",
  definition.uri = "",
  scaling.factor = "",
  quantity.kind = "",
  si.scaling.factor = ""
)
}
\arguments{
\item{label}{A string. Label describing the Unit.}

\item{description}{A string. A explanation of what the Unit represents.}

\item{from.existing}{A string. URI of an existing Unit to reuse. Can use built-in list Units to auto-fill.}

\item{definition.uri}{A string. Source URI for the definition of the Unit}

\item{scaling.factor}{A number. Scaling factor of the observation to base unit.}

\item{quantity.kind}{A string. The QUDT quantity kind this unit is used to measure. Can use built-in list QuantityKinds to auto-fill.}

\item{si.scaling.factor}{A number. Scaling factor to convert from this unit to the SI base unit defined in the unit's quantity_kind.}
}
\description{
Create a unit object for add.unit.column() or add.observation.column()
}
\examples{
unit.value(
  label = "meter", description = "meter",
  from.existing = Units$M, si.scaling.factor = 1,
  quantity.kind = QuantityKinds$Length
)
}
