% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{unit.values}
\alias{unit.values}
\title{Create a unit.values object.}
\usage{
unit.values(...)
}
\arguments{
\item{...}{unit objects created by unit.value() or strings.}
}
\description{
Create a unit.values object for use in other functions
}
\examples{
data(example.data)
config <- create.config(example.data, title = "sweden_at_eurovision_no_missing")
config <- add.unit.column(config,
  column.name = "Value",
  unit.values = unit.values(
    unit.value(label = "Unitless"),
    unit.value(label = "Number")
  )
)
}
