% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_value_from_existing.R
\docType{data}
\name{Units}
\alias{Units}
\title{A list of strings that can be used to fill in the from.existing argument in unit.value().}
\format{
An object of class \code{list} of length 1727.
}
\usage{
Units
}
\description{
A list of strings that can be used to fill in the from.existing argument in unit.value().
}
\keyword{datasets}
