% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{create.config}
\alias{create.config}
\title{creating the configuration}
\usage{
create.config(
  df,
  id = "",
  title = "",
  description = "",
  summary = "",
  publisher = "",
  creator = "",
  themes = "",
  keywords = "",
  dataset.issued = "",
  dataset.modified = "",
  license = "",
  public.contact.point.uri = ""
)
}
\arguments{
\item{df}{A dataframe. The dataframe should be in the standard shape, refer to https://gss-cogs.github.io/csvcubed-docs/external/guides/shape-data/#standard-shape.}

\item{id}{A string. A unique id for the cube.}

\item{title}{A string. A title for this data set.}

\item{description}{A string. A longer description of the cube.}

\item{summary}{A string. A shorter summary of the data set.}

\item{publisher}{A string. The publisher of the data set (uri). Can use built-in list "Organizations" to auto-fill.}

\item{creator}{A string. The creator of the data set (uri). Can use built-in list "Organizations" to auto-fill.}

\item{themes}{A vector of strings. The themes of the data set.}

\item{keywords}{A vector of strings. The keywords of the dataset.}

\item{dataset.issued}{A String. Date time that the data set was initially published in ISO 8601 format, e.g. 2022-03-31 or 2022-03-31T12:54:30Z.}

\item{dataset.modified}{A String. Date time that the data set was last modified published in ISO 8601 format, e.g. 2022-03-31 or 2022-03-31T12:54:30Z.}

\item{license}{A String. URI that represents the copyright license applying to this cube. Can use built-in list "Licences" to auto-fill.}

\item{public.contact.point.uri}{A string. URI that provides a public contact point for discussion of the data set, e.g. mailto:contact.point@example.com.}
}
\value{
A configuration object.
}
\description{
Users can use this function to configure the metadata of the configuration, and then add column configurations by using other functions.
}
\examples{
# load data
data(example.data)
# create a configuration for example.data
config <- create.config(example.data,
  title = "Sweden at Eurovision",
  summary = "List of Swedish entries to the Eurovision Song Contest since 1958.",
  license = "https://creativecommons.org/licenses/by/4.0/",
  publisher = Organizations$Open_Knowledge_Foundation,
  dataset.issued = "2022-04-08",
  keywords = c("Eurovision", "Song Contest", "Sweden", "European Broadcasting Union")
)
}
