% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generate.json.configuration}
\alias{generate.json.configuration}
\title{Generate the configuration json file}
\usage{
generate.json.configuration(config, file)
}
\arguments{
\item{config}{An configuration object. The configuration to be serialized into json.}

\item{file}{A string. The directory path and file name of the json file.}
}
\value{
The json scheme string of the  configuration.
}
\description{
Generate the configuration json file to the working diretory. Returns the json string.
}
\examples{
\dontrun{
data(example.data)
config <- create.config(
  df = example.data,
  id = "12341241", title = "sweden_at_eurovision_no_missing",
  dataset.issued = "2022-04-08",
  keywords = c("Eurovision", "Song Contest", "Sweden", "European Broadcasting Union")
)

# the following will return the json string and generate a json file
# called "my-data.json" in the current working directory.
generate.json.configuration(config, "my-data.json")

# Make sure your csv file has the same name
write.csv(example.data, "my-data.csv")
}
}
